Heavy refactoring the way that we manage relations between users and promises
Now we have a template tag with the current 3 possible states: owner, supporter and
nothing. You can use it in every context that provides a user and a promise,
like home.html and promise.html.

Also, the home view does not require signing in any more.

ps.: this commit breaks the home page for anonymous users, cause it renders the
support url, that currently needs an user id. This will be changed in the next
commit.